fix: 将插件app改名为application
