+++ application/application_global.h
@@ -1,12 +1,12 @@
-#ifndef APP_GLOBAL_H
-#define APP_GLOBAL_H
+#ifndef APPLICATION_GLOBAL_H
+#define APPLICATION_GLOBAL_H
 
 #include <QtCore/qglobal.h>
 
-#if defined(APP_LIBRARY)
+#if defined(APPLICATION_LIBRARY)
 #  define APPSHARED_EXPORT Q_DECL_EXPORT
 #else
 #  define APPSHARED_EXPORT Q_DECL_IMPORT
 #endif
 
-#endif // APP_GLOBAL_H
+#endif // APPLICATION_GLOBAL_H
