+++ application/plugin.json
@@ -1,5 +1,5 @@
 {
-  "Name": "App",
+  "Name": "Application",
   "Version": "1.0.0",
   "CompatVersion":"5.3",
   "Copyright": "(C) 2014-2020 syberos.com",
